styles: #6 게시글 화면, 커뮤니티 화면 스타일 분리
