+++ src/components/article/ArticleTitle.tsx
@@ -1,41 +1,20 @@
-import React from 'react';
-import {View, StyleSheet, Platform, Text} from 'react-native';
+import React, {Component} from 'react';
+import {View, Text} from 'react-native';
+import styles from '../../styles/article/ArticleTitle';
 
-const ArticleTitle = () => {
-    const styles = StyleSheet.create({
-        titleContainer: {
-            backgroundColor: 'white'
-        },
-
-        titleText: {
-            marginTop: Platform.OS == 'ios' ? 10 : 5,
-            marginLeft: 20,
-            marginRight: 10,
-            color: 'black',
-            fontSize: Platform.OS == 'ios' ? 17 : 14,
-            fontFamily: 'GmarketSansTTFBold',
-            lineHeight: 20
-        },
-
-        subText: {
-            marginTop: Platform.OS == 'ios' ? 10 : 5,
-            marginLeft: 20,
-            color: 'gray',
-            fontSize: Platform.OS == 'ios' ? 15 : 12,
-            fontFamily: 'GmarketSansTTFMedium'
-        }
-    });
-
-    return (
-        <View style = {styles.titleContainer}>
-            <Text style = {styles.titleText}>
-                How to Create a Telegram Bot Using Python Making $300 Per Month
-            </Text>
-            <Text style = {styles.subText}>
-                A Step-by-Step Guide
-            </Text>
-        </View>
-    );
+class ArticleTitle extends Component {
+    render() {
+        return (
+            <View style = {styles.titleContainer}>
+                <Text style = {styles.titleText}>
+                    How to Create a Telegram Bot Using Python Making $300 Per Month
+                </Text>
+                <Text style = {styles.subText}>
+                    A Step-by-Step Guide
+                </Text>
+            </View>
+        )
+    };
 };
 
 export default ArticleTitle;
