+++ src/components/community/Searchbox.tsx
@@ -1,23 +1,17 @@
-import React from 'react';
-import {TextInput, StyleSheet} from 'react-native';
+import React, {Component} from 'react';
+import {TextInput} from 'react-native';
+import styles from '../../styles/community/SearchBox';
 
-const Searchbox = () => {
-    const styles = StyleSheet.create({
-        inputStyle: {
-            padding: 10,
-            backgroundColor: 'white',
-            borderColor: '#FFFFF',
-            borderRadius: 8
-        }
-    });
-
-    return (
-        <TextInput 
-            style = {styles.inputStyle}
-            placeholder = '사람, 태그, 키워드로 검색하기'
-        >
-        </TextInput>
-    );
+class SearchBox extends Component {
+    render(){
+        return (
+            <TextInput 
+                style = {styles.inputStyle}
+                placeholder = '사람, 태그, 키워드로 검색하기'
+            >
+            </TextInput>
+        );
+    };
 };
 
-export default Searchbox;
+export default SearchBox;
