+++ src/styles/community/SearchBox.js
@@ -1,0 +1,12 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    inputStyle: {
+        padding: 10,
+        backgroundColor: 'white',
+        borderColor: '#FFFFF',
+        borderRadius: 8
+    }
+});
+
+export default styles;
