+++ src/styles/community/SearchButton.js
@@ -1,0 +1,14 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    buttonStyle: {
+        marginTop: 10,
+        marginLeft: 20,
+        marginBottom: 10,
+        padding: 10,
+        backgroundColor: '#2BC63B',
+        borderRadius: 8
+    }
+});
+
+export default styles;
